pyproject.toml
[tool.poetry]
name = "vending-machine"
version = "0.1.0"
description = "Requirements for the Vending Machine API"
authors = ["Mohammad Kabir Abdulsalam <mkab.abdulsalam@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
django-extensions = "^3.1.5"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
pytest = "^7.0.1"
pytest-django = "^4.5.2"
flake8 = "^4.0.1"
black = "^22.1.0"
isort = "^5.10.1"

[tool.black]
line-length=99
target-version = ['py37']

[tool.isort]
multi_line_output = 3
line_length = 99
use_parentheses = true
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
